refactor: :recycle: merge search module and controller in application ressource
