+++ backend/src/product/application.module.ts
@@ -6,7 +6,7 @@ import { ExportApplicationsUseCase } from './application/usecases/application-ex
 import { ApplicationRepository } from './infrastructure/repository/application.repository';
 
 import { CommonModule } from 'src/common/common.module';
-import { PrismaModule } from 'src/prisma/prisma.module'; // ✅ ici !
+import { PrismaModule } from 'src/prisma/prisma.module';
 import { LabelsModule } from 'src/labels/labels.module';
 import { MetadatasModule } from 'src/metadatas/metadatas.module';
