--- backend/src/search/infrastructure/search.repository.interface.ts
@@ -1,8 +1,0 @@
-import { SearchApplicationDto } from '../application/dto/search-application.dto';
-
-export interface IApplicationSearchRepository {
-  findApplicationsBySearch(dto: SearchApplicationDto): Promise<{
-    results: any[];
-    total: number;
-  }>;
-}
